fix issue#2323
