+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-sentinel/src/main/java/com/alibaba/cloud/sentinel/SentinelWebAutoConfiguration.java
@@ -36,7 +36,6 @@ import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.util.StringUtils;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
 /**
@@ -64,23 +63,6 @@ public class SentinelWebAutoConfiguration implements WebMvcConfigurer {
 
 	@Autowired
 	private Optional<RequestOriginParser> requestOriginParserOptional;
-
-	@Autowired
-	private Optional<SentinelWebInterceptor> sentinelWebInterceptorOptional;
-
-	@Override
-	public void addInterceptors(InterceptorRegistry registry) {
-		if (!sentinelWebInterceptorOptional.isPresent()) {
-			return;
-		}
-		SentinelProperties.Filter filterConfig = properties.getFilter();
-		registry.addInterceptor(sentinelWebInterceptorOptional.get())
-				.order(filterConfig.getOrder())
-				.addPathPatterns(filterConfig.getUrlPatterns());
-		log.info(
-				"[Sentinel Starter] register SentinelWebInterceptor with urlPatterns: {}.",
-				filterConfig.getUrlPatterns());
-	}
 
 	@Bean
 	@ConditionalOnProperty(name = "spring.cloud.sentinel.filter.enabled",
@@ -116,6 +98,13 @@ public class SentinelWebAutoConfiguration implements WebMvcConfigurer {
 		urlCleanerOptional.ifPresent(sentinelWebMvcConfig::setUrlCleaner);
 		requestOriginParserOptional.ifPresent(sentinelWebMvcConfig::setOriginParser);
 		return sentinelWebMvcConfig;
+	}
+
+	@Bean
+	@ConditionalOnProperty(name = "spring.cloud.sentinel.filter.enabled",
+			matchIfMissing = true)
+	public SentinelWebMvcConfigurer sentinelWebMvcConfigurer() {
+		return new SentinelWebMvcConfigurer();
 	}
 
 }
